app/requirements.txt
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.5.0
beautifulsoup4==4.10.0
bidict==0.21.4
blinker==1.4
Brlapi==0.7.0
build-flask-app==0.1.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
command-not-found==0.3
configobj==5.0.6
cryptography==36.0.1
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
distlib==0.3.4
distro==1.7.0
dnspython==2.2.0
docopt==0.6.2
docx2txt==0.8
email-validator==1.1.3
entrypoints==0.4
eventlet==0.33.0
filelock==3.6.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
fonttools==4.29.1
futures==3.0.5
gevent==21.12.0
goslate==1.5.2
greenlet==1.1.2
grpcio==1.44.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
ifaddr==0.1.7
importlib-metadata==4.11.2
itsdangerous==2.1.0
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.1.0
keyring==23.5.0
kiwisolver==1.3.2
langdetect==1.0.9
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.6
MarkupSafe==2.1.0
matplotlib==3.5.1
nemo-emblems==4.6.0
netifaces==0.11.0
nltk==3.7
nose==1.3.7
numpy==1.22.2
oauthlib==3.2.0
onboard==1.4.1
packaging==21.3
PAM==0.4.2
pbr==5.8.1
pdf2image==1.16.0
pdfminer==20191125
pdfminer.six==20211012
pexpect==4.8.0
Pillow==9.0.1
pipreqs==0.4.11
platformdirs==2.5.1
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.14.1
pycups==1.9.73
pycurl==7.43.0.2
pyforest==1.1.0
Pygments==2.11.2
PyGObject==3.36.0
PyICU==2.4.2
pyinotify==0.9.6
PyInquirer==1.0.3
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.0.1
PyNaCl==1.5.0
pyparsing==3.0.7
pyparted==3.11.2
pyRFC3339==1.1
python-apt==2.0.0
python-dateutil==2.8.2
python-debian==0.1.43
python-dotenv==0.19.2
python-engineio==4.3.1
python-socketio==5.5.2
python-xapp==2.0.1
python-xlib==0.31
pytz==2021.3
pyxdg==0.27
PyYAML==6.0
regex==2022.1.18
reportlab==3.6.7
requests==2.27.1
requests-unixsocket==0.3.0
SecretStorage==3.3.1
setproctitle==1.2.2
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
systemd-python==234
termcolor==1.1.0
testresources==2.0.1
tinycss2==1.1.1
tqdm==4.63.0
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==1.3.3
urllib3==1.26.8
validators==0.18.2
virtualenv==20.13.2
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
WTForms==3.0.1
xkit==0.0.0
yarg==0.1.9
yolk==0.4.3
zeroconf==0.38.4
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
